Fix BIN export & use proper BIN extract path
Was producing invalid BINs before since I forgot to add image count to beginning of file. Was also missing the "b" at the start of the filename. Exporting was overwriting the contents of the image folder rather than creating a new "Extracted" folder at the location of the input BIN.