Improve export performance by avoiding BIN dupes
Avoids exporting the same BIN multiple times if the same image is use. Uses a memorystream instead of a filestream so the byte array of BINs with a unique base image get added to a list upon creation. BIN creation is skipped if one with a matching base image already exist, and the existing byte array is used for output instead.